Refactor functions GUI
